Refactored location of DI classes
